(22/10) Proteção de rotas no back-end e no front-end

+++ back-end/src/controllers/users.js
@@ -6,6 +6,11 @@ const controller = {}     // Objeto vazio
 
 controller.create = async function(req, res) {
   try {
+
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
     
     if (req.body.password) {
       
@@ -28,6 +33,11 @@ controller.create = async function(req, res) {
 
 controller.retrieveAll = async function(req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
+
+
     const result = await prisma.user.findMany({
       // Omite o campo "password" do resultado
       // por questão de segurança
@@ -47,6 +57,15 @@ controller.retrieveAll = async function(req, res) {
 
 controller.retrieveOne = async function(req, res) {
   try {
+
+    // Somente usuários administradores ou o próprio usuário
+    // autenticado podem acessar este recurso
+    // HTTP 403: Forbidden
+    if(! (req?.authUser?.is_admin || 
+      Number(req?.authUser?.id) === Number(req.params.id))) 
+      return res.status(403).end()
+
+
     const result = await prisma.user.findUnique({
       // Omite o campo "password" do resultado
       // por questão de segurança
@@ -69,6 +88,9 @@ controller.retrieveOne = async function(req, res) {
 
 controller.update = async function(req, res) {
   try {
+    // Somente usuários administradores podem acessar este recurso
+    // HTTP 403: Forbidden(
+    if(! req?.authUser?.is_admin) return res.status(403).end()
 
     if(req.body.password) {
       
@@ -95,6 +117,13 @@ controller.update = async function(req, res) {
 }
 
 controller.delete = async function(req, res) {
+  
+  // Somente usuários administradores podem acessar este recurso
+  // HTTP 403: Forbidden
+  if(! req?.authUser?.is_admin) {
+    return res.status(403).end()
+  }
+  
   try {
     await prisma.user.delete({
       where: { id: Number(req.params.id) }
